Fixed bug caused by empty factor levels
If a factor covariate in the dfunc has 3 levels, but only 2 are present in the data (e.g., while bootstrapping), the dfunc does not estimate a coef for the missing level.  However, predict.dfunc assumes all levels in the input sites data have a coefficient in the dfunc.  This fix lets the missing level stay missing.